requirements.txt
nltk==3.9.1
peft==0.14.0
torch==2.2.2
numpy==1.26.4
pandas==2.2.3
emoji==2.14.1
pillow==11.0.0
datasets==3.2.0
selenium==4.29.0
pydrive2==1.21.3
requests==2.32.3
tensorboard==2.19.0
torchvision==0.21.0
torchvision==0.17.2
transformers==4.47.1
tensorboardX==2.6.2.2
beautifulsoup4==4.13.3
webdriver-manager==4.0.2
opencv-python==4.11.0.86